refactor: moved cart reducer to reducers folder
