+++ src/reducers/cart-reducer.js
@@ -1,0 +1,10 @@
+const cartReducer = (cartState, action) => {
+  switch (action.type) {
+    case "cartList":
+      return { ...cartState, cartList: action.payload };
+    default:
+      return cartState;
+  }
+};
+
+export { cartReducer };
