+++ src/reducers/index.js
@@ -1,1 +1,2 @@
 export { productPageReducer } from "./product-page-reducer";
+export { cartReducer } from "./cart-reducer";
